feat(socket): pubsub compression implementation
