+++ msg-socket/src/sub/driver.rs
@@ -1,6 +1,7 @@
 use bytes::Bytes;
 use futures::{Future, Stream, StreamExt};
 use msg_common::unix_micros;
+use msg_wire::compression::Compressor;
 use std::collections::{HashSet, VecDeque};
 use std::net::SocketAddr;
 use std::pin::Pin;
@@ -34,6 +35,8 @@ pub(crate) struct SubDriver<T: ClientTransport> {
     pub(super) to_socket: mpsc::Sender<PubMessage>,
     /// A joinset of authentication tasks.
     pub(super) connection_tasks: JoinSet<ConnectionResult<T::Io, T::Error>>,
+    /// Optional payload (de)compressor.
+    pub(super) compressor: Option<Arc<dyn Compressor>>,
     /// The set of subscribed topics.
     pub(super) subscribed_topics: HashSet<String>,
     /// All active publisher sessions for this subscriber socket.
@@ -55,7 +58,19 @@ where
         loop {
             if let Poll::Ready(Some((addr, result))) = this.publishers.poll_next_unpin(cx) {
                 match result {
-                    Ok(msg) => {
+                    Ok(mut msg) => {
+                        if let Some(ref compressor) = this.compressor {
+                            let Ok(decompressed) = compressor.decompress(&msg.payload) else {
+                                error!(
+                                    topic = msg.topic.as_str(),
+                                    "Failed to decompress message payload"
+                                );
+
+                                continue;
+                            };
+
+                            msg.payload = decompressed;
+                        }
                         this.on_message(PubMessage::new(addr, msg.topic, msg.payload));
                     }
                     Err(e) => {
@@ -94,6 +109,10 @@ impl<T> SubDriver<T>
 where
     T: ClientTransport + Send + Sync + 'static,
 {
+    pub fn set_compressor<C: Compressor>(&mut self, compressor: C) {
+        self.compressor = Some(Arc::new(compressor));
+    }
+
     fn on_command(&mut self, cmd: Command) {
         debug!("Received command: {:?}", cmd);
         match cmd {
@@ -121,7 +140,6 @@ where
                 let id = self.options.auth_token.clone();
                 let transport = Arc::clone(&self.transport);
 
-                // NOTE: don't know if this is gonna work
                 self.connection_tasks.spawn(async move {
                     let io = transport.connect_with_auth(endpoint, id).await?;
