+++ msg/examples/pubsub_compression.rs
@@ -4,31 +4,28 @@ use std::time::Duration;
 use tokio::time::timeout;
 use tracing::Instrument;
 
-use msg::{PubOptions, PubSocket, SubOptions, SubSocket, Tcp, TcpOptions};
+use msg::{compression::GzipCompressor, PubSocket, SubSocket, Tcp, TcpOptions};
 
 #[tokio::main]
 async fn main() {
     let _ = tracing_subscriber::fmt::try_init();
     // Configure the publisher socket with options
-    let mut pub_socket = PubSocket::with_options(
-        Tcp::new(),
-        PubOptions::default()
-            .backpressure_boundary(8192)
-            .session_buffer_size(1024)
-            .flush_interval(Duration::from_micros(100)),
-    );
+    let mut pub_socket = PubSocket::new(Tcp::new())
+        // Enable Gzip compression (compression level 6)
+        .with_compressor(GzipCompressor::new(6));
 
     // Configure the subscribers with options
-    let mut sub1 = SubSocket::with_options(
+    let mut sub1 = SubSocket::new(
         // TCP transport with blocking connect, usually connection happens in the background.
         Tcp::new_with_options(TcpOptions::default().with_blocking_connect()),
-        SubOptions::default().ingress_buffer_size(1024),
-    );
+    )
+    // Enable Gzip decompression (at the same level)
+    .with_compressor(GzipCompressor::new(6));
 
-    let mut sub2 = SubSocket::with_options(
-        Tcp::new_with_options(TcpOptions::default().with_blocking_connect()),
-        SubOptions::default().ingress_buffer_size(1024),
-    );
+    let mut sub2 = SubSocket::new(Tcp::new_with_options(
+        TcpOptions::default().with_blocking_connect(),
+    ))
+    .with_compressor(GzipCompressor::new(6));
 
     tracing::info!("Setting up the sockets...");
     pub_socket.bind("127.0.0.1:0").await.unwrap();
