+++ msg/src/lib.rs
@@ -1,2 +1,3 @@
 pub use msg_socket::*;
 pub use msg_transport::*;
+pub use msg_wire::compression;
